chore: 디렉터리 정리
