+++ src/app/posts/contents/[...slug]/page.tsx
@@ -1,6 +1,6 @@
 import { type Metadata } from "next";
 import { notFound } from "next/navigation";
-import PostArticleContent from "@/components/post-article/post-article-content";
+import PostArticleContent from "@/components/post-article/content/post-article-content";
 import PostArticleHeader from "@/components/post-article/post-article-header";
 import PostArticleTOCSidebar from "@/components/post-article/toc/post-article-toc-sidebar";
 import HorizontalSeparator from "@/components/separator/horizontal-separator";
